Fix #123716: autokeying Shape Key fields does not work
The issue had essentially the same origin as #123538 (fixed
in #123719), but in the autokeying code. The autokeying code was
already fully resolving the RNA path to be relative to the owning
ID, but the RNA pointer getting passed with that fully-resolved path
to the keying code was the struct for the not-fully-resolved path.

This fixes it by simply passing the RNA pointer for the owning ID
rather than the struct being keyed.

Pull Request: https://projects.blender.org/blender/blender/pulls/123721